+++ src/DiffEditor/index.ts
@@ -1,5 +1,7 @@
-import { memo } from 'react';
+import { memo } from "react";
 
-import DiffEditor from './DiffEditor';
+import DiffEditor from "./DiffEditor";
+
+export * from "./types";
 
 export default memo(DiffEditor);
